+++ Documents/NetBeansProjects/WebshopJavaApplicatie/klant.xml
@@ -1,17 +1,11 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<java version="1.8.0_73" class="java.beans.XMLDecoder">
- <object class="MAIN.KlantXML">
-  <void property="achternaam">
-   <string>d</string>
-  </void>
-  <void property="email">
-   <string>ggggg</string>
-  </void>
-  <void property="tussenvoegsel">
-   <string>s</string>
-  </void>
-  <void property="voornaam">
-   <string>a</string>
-  </void>
- </object>
-</java>
+<object-stream>
+  <string>&lt;klanten&gt;
+  &lt;klant&gt;
+    &lt;klantId&gt;1&lt;/klantId&gt;
+    &lt;voornaam&gt;h&lt;/voornaam&gt;
+    &lt;achternaam&gt;w&lt;/achternaam&gt;
+    &lt;tussenvoegsel&gt;j&lt;/tussenvoegsel&gt;
+    &lt;email&gt;ret@me.be&lt;/email&gt;
+  &lt;/klant&gt;
+&lt;/klanten&gt;</string>
+</object-stream>
