--- Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/Controller/ControllerInterface.java
@@ -1,23 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Controller;
-
-/**
- *
- * @author Excen
- */
-public interface ControllerInterface {
-    
-    // Alle methoden die we nodig hebben in de controller. De controller krijgt input van de view (in de vorm dat de 
-    // view alle methoden kan aanroepen) en spreek zelf de implementatie van de verschillende DAO methoden aan om de mogelijke 
-    // taken te doen.
-    
-    // Checken of klant als bestaat aan de hand van klantId
-    
-    
-    // 
-    
-}
