+++ Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/DAOs/Impl/FireBird/BestellingDAOFB.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package DAOs.Impl;
+package DAOs.Impl.FireBird;
 
+import DAOs.Impl.MySQL.BestellingArtikelDAOSQL;
+import DAOs.Impl.MySQL.BestellingDAOSQL;
 import DAOs.Interface.BestellingDAOInterface;
 import POJO.Bestelling;
 import java.sql.Connection;
