+++ Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/DAOs/Impl/MySQL/BestellingArtikelDAOSQL.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package DAOs.Impl;
+package DAOs.Impl.MySQL;
 
 import ConnectionPool.HikariCP;
 import DAOs.Interface.BestellingDAOInterface;
