+++ Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/DAOs/Interface/AdresDAOInterface.java
@@ -7,6 +7,8 @@ package DAOs.Interface;
 import POJO.Adres;
 import java.util.ArrayList;
 
+
+
 /**
  *
  * @author Wendy
