+++ Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/Factory/DaoFactory.java
@@ -6,20 +6,20 @@
 package Factory;
 
 import Controller.HoofdMenuController;
-import DAOs.Impl.AdresDAOFB;
-import DAOs.Impl.AdresDAOSQL;
-import DAOs.Impl.ArtikelDAOFB;
-import DAOs.Impl.ArtikelDAOSQL;
-import DAOs.Impl.BestellingArtikelDAOFB;
-import DAOs.Impl.BestellingArtikelDAOSQL;
-import DAOs.Impl.BestellingDAOFB;
-import DAOs.Impl.BestellingDAOSQL;
-import DAOs.Impl.KlantAdresDAOFB;
-import DAOs.Impl.KlantAdresDAOSQL;
-import DAOs.Impl.KlantDAOFB;
-import DAOs.Impl.KlantDAOJSON;
-import DAOs.Impl.KlantDAOSQL;
-import DAOs.Impl.KlantDAOXML;
+import DAOs.Impl.FireBird.AdresDAOFB;
+import DAOs.Impl.MySQL.AdresDAOSQL;
+import DAOs.Impl.FireBird.ArtikelDAOFB;
+import DAOs.Impl.MySQL.ArtikelDAOSQL;
+import DAOs.Impl.FireBird.BestellingArtikelDAOFB;
+import DAOs.Impl.MySQL.BestellingArtikelDAOSQL;
+import DAOs.Impl.FireBird.BestellingDAOFB;
+import DAOs.Impl.MySQL.BestellingDAOSQL;
+import DAOs.Impl.FireBird.KlantAdresDAOFB;
+import DAOs.Impl.MySQL.KlantAdresDAOSQL;
+import DAOs.Impl.FireBird.KlantDAOFB;
+import DAOs.Impl.Json_XML.KlantDAOJSON;
+import DAOs.Impl.MySQL.KlantDAOSQL;
+import DAOs.Impl.Json_XML.KlantDAOXML;
 import DAOs.Interface.AdresDAOInterface;
 import DAOs.Interface.ArtikelDAOInterface;
 import DAOs.Interface.BestellingArtikelDAOInterface;
