+++ Documents/NetBeansProjects/WebshopJavaApplicatie/src/main/java/MAIN/Main.java
@@ -3,9 +3,9 @@ package MAIN;
 import Controller.HoofdMenuController;
 import Controller.KlantController;
 import Controller.ArtikelController;
-import DAOs.Impl.AdresDAOSQL;
-import DAOs.Impl.ArtikelDAOSQL;
-import DAOs.Impl.KlantDAOSQL;
+import DAOs.Impl.MySQL.AdresDAOSQL;
+import DAOs.Impl.MySQL.ArtikelDAOSQL;
+import DAOs.Impl.MySQL.KlantDAOSQL;
 import DAOs.Interface.AdresDAOInterface;
 import DAOs.Interface.KlantDAOInterface;
 import POJO.Adres;
@@ -20,15 +20,19 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
-
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 /**
  *
  * @author Excen
  */
 
-public class Main {       
+public class Main {     
+    private static final Logger _log = LoggerFactory.getLogger(Main.class);
     
     public static void main (String[]args) throws SQLException, ClassNotFoundException, FileNotFoundException{
+    
+        
         
         String driver = "com.mysql.jdbc.Driver";
         String url = "jdbc:mysql://localhost:3306/winkel?autoReconnect=true&useSSL=false";
@@ -45,7 +49,7 @@ public class Main {
         
         
         HoofdMenuController start = new HoofdMenuController();
-            start.start(); 
+            start.setDatabase();
             
       
         /*
